Docker file updates + Spotify Auth Updates

+++ .next/server/chunks/[root-of-the-server]__8cbe5c35._.js
@@ -128,9 +128,12 @@ var __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$
 ;
 const playlist_id = '1w7opBRG814H7CMZaMOCN7';
 async function add_track_to_playlist() {
+    const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
+    const url = new URL(`https://api.spotify.com/v1/playlists/${playlist_id}/tracks`);
     try {
-        const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
-        const url = new URL(`https://api.spotify.com/v1/playlists/${playlist_id}/tracks`);
+        if (!token_data) {
+            throw new Error("Unable to get access token from file");
+        }
         const res = await fetch(url, {
             method: 'POST',
             headers: {
@@ -155,9 +158,12 @@ async function add_track_to_playlist() {
     }
 }
 async function remove_track_from_playlist() {
+    const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
+    const url = new URL(`https://api.spotify.com/v1/playlists/${playlist_id}/tracks`);
     try {
-        const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
-        const url = new URL(`https://api.spotify.com/v1/playlists/${playlist_id}/tracks`);
+        if (!token_data) {
+            throw new Error("Unable to get access token from file");
+        }
         const res = await fetch(url, {
             method: 'DELETE',
             headers: {
@@ -195,6 +201,9 @@ async function search_spotify() {
     url.searchParams.append('market', market);
     url.searchParams.append('limit', limit);
     try {
+        if (!token_data) {
+            throw new Error("Unable to get access token from file");
+        }
         const res = await fetch(url, {
             method: 'GET',
             headers: {
@@ -213,11 +222,14 @@ async function search_spotify() {
     }
 }
 async function get_playlist() {
+    const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
+    const market = 'US';
+    const url = new URL(`https://api.spotify.com/v1/playlists/${playlist_id}`);
+    url.searchParams.append('market', market);
     try {
-        const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
-        const market = 'US';
-        const url = new URL(`https://api.spotify.com/v1/playlists/${playlist_id}`);
-        url.searchParams.append('market', market);
+        if (!token_data) {
+            throw new Error("Unable to get access token from file");
+        }
         const res = await fetch(url, {
             method: 'GET',
             headers: {
@@ -243,11 +255,14 @@ async function get_playlist() {
     }
 }
 async function get_current_track() {
+    const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
+    const market = 'US';
+    const url = new URL('https://api.spotify.com/v1/me/player/currently-playing');
+    url.searchParams.append('market', market);
     try {
-        const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
-        const market = 'US';
-        const url = new URL('https://api.spotify.com/v1/me/player/currently-playing');
-        url.searchParams.append('market', market);
+        if (!token_data) {
+            throw new Error("Unable to get access token from file");
+        }
         const res = await fetch(url.toString(), {
             method: 'GET',
             headers: {
@@ -294,12 +309,15 @@ async function extract_data_curr_song(data) {
     };
 }
 async function get_top_items() {
+    const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
+    const url = new URL(`https://api.spotify.com/v1/me/top/tracks`);
+    url.searchParams.append("type", "tracks");
+    url.searchParams.append("time_range", "medium_term");
+    url.searchParams.append("limit", "10");
     try {
-        const token_data = await (0, __TURBOPACK__imported__module__$5b$project$5d2f$src$2f$app$2f$spotify_utils$2f$json_actions$2e$ts__$5b$app$2d$route$5d$__$28$ecmascript$29$__["get_local_token"])();
-        const url = new URL(`https://api.spotify.com/v1/me/top/tracks`);
-        url.searchParams.append("type", "tracks");
-        url.searchParams.append("time_range", "medium_term");
-        url.searchParams.append("limit", "10");
+        if (!token_data) {
+            throw new Error("Unable to get access token from file");
+        }
         const res = await fetch(url, {
             method: "GET",
             headers: {
@@ -324,7 +342,7 @@ async function get_top_items() {
     }
 }
 (async function main() {
-//console.log(await get_top_items())
+    console.log(await get_top_items());
 //console.log(await get_current_track())
 //console.log(await get_playlist())
 //console.log(await search_spotify())
